Added tests for Security Service. Refactored tests for User Controller. Autowired UserServiceInterceptor.

+++ src/main/java/bar/dao/EmployeeRoleDAO.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import bar.model.EmployeeRole;
 
 @Repository
-public interface EmployeeRoleDAO extends CrudRepository<EmployeeRole, Byte> {
+public interface EmployeeRoleDAO extends CrudRepository<EmployeeRole, Long> {
 	EmployeeRole findByName(String name);
 
 	boolean existsByName(String name);
